requirements.txt
opencv-python>=4.8.0
mediapipe-aarch64>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
streamlit>=1.25.0
plotly>=5.15.0
scikit-learn>=1.1.0
scipy>=1.9.0
Pillow>=9.0.0
tqdm>=4.64.0 